- split NoteEdit into multiple files (component, container, index) - fix keyboard overlap problem on ios
