+++ app/containers/NoteEdit/index.js
@@ -1,0 +1,1 @@
+export {NodeEditContainer as default} from './container'
